moved role_rule table to common
