+++ vantage6-algorithm-store/vantage6/algorithm/store/db.py
@@ -15,6 +15,7 @@ from vantage6.algorithm.store.model import (
     User,
     Review,
     Vantage6Server,
-    role_rule_association,
     Policy,
 )
+
+from vantage6.backend.common.permission_models import role_rule_association
