+++ vantage6-algorithm-store/vantage6/algorithm/store/model/__init__.py
@@ -7,7 +7,6 @@ from vantage6.algorithm.store.model.function import Function
 from vantage6.algorithm.store.model.role import Role
 from vantage6.algorithm.store.model.rule import Rule
 from vantage6.algorithm.store.model.vantage6_server import Vantage6Server
-from vantage6.algorithm.store.model.role_rule_association import role_rule_association
 from vantage6.algorithm.store.model.permission import Permission
 from vantage6.algorithm.store.model.user import User
 from vantage6.algorithm.store.model.developer_algorithm_association import (
@@ -16,3 +15,4 @@ from vantage6.algorithm.store.model.developer_algorithm_association import (
 from vantage6.algorithm.store.model.review import Review
 from vantage6.algorithm.store.model.ui_visualization import UIVisualization
 from vantage6.algorithm.store.model.policy import Policy
+from vantage6.backend.common.permission_models import role_rule_association
