+++ vantage6-algorithm-store/vantage6/algorithm/store/model/rule.py
@@ -5,7 +5,7 @@ from sqlalchemy import Column, Text, String
 from sqlalchemy.orm import relationship
 from sqlalchemy.orm.exc import NoResultFound
 from vantage6.backend.common.base import Base, DatabaseSessionManager
-from vantage6.backend.common.permission_interfaces import RuleInterface
+from vantage6.backend.common.permission_models import RuleInterface
 
 
 class Operation(str, Enumerate):
