+++ vantage6-backend-common/vantage6/backend/common/permission_models.py
@@ -3,8 +3,16 @@ from __future__ import annotations
 from typing import Any
 
 from sqlalchemy.orm.exc import NoResultFound
+from sqlalchemy import Column, Integer, ForeignKey, Table
 
-from vantage6.backend.common.base import DatabaseSessionManager
+from vantage6.backend.common.base import Base, DatabaseSessionManager
+
+role_rule_association = Table(
+    "role_rule_association",
+    Base.metadata,
+    Column("role_id", Integer, ForeignKey("role.id")),
+    Column("rule_id", Integer, ForeignKey("rule.id")),
+)
 
 
 class RuleInterface:
@@ -54,3 +62,4 @@ class RoleInterface:
             String representation of the role
         """
         raise NotImplemented("__repr__ method must be implemented for Role class")
+
