+++ vantage6-server/vantage6/server/model/__init__.py
@@ -1,8 +1,9 @@
+from vantage6.backend.common.base import Base
+from vantage6.backend.common.permission_models import role_rule_association
 from vantage6.server.model.authenticatable import Authenticatable
 from vantage6.server.model.user import User
 from vantage6.server.model.organization import Organization
 from vantage6.server.model.collaboration import Collaboration
-from vantage6.backend.common.base import Base
 from vantage6.server.model.member import Member, StudyMember
 from vantage6.server.model.node_config import NodeConfig
 from vantage6.server.model.node import Node
@@ -11,7 +12,6 @@ from vantage6.server.model.run import Run
 from vantage6.server.model.permission import Permission, UserPermission
 from vantage6.server.model.role import Role
 from vantage6.server.model.rule import Rule
-from vantage6.server.model.role_rule_association import role_rule_association
 from vantage6.server.model.algorithm_port import AlgorithmPort
 from vantage6.server.model.task_database import TaskDatabase
 from vantage6.server.model.algorithm_store import AlgorithmStore
